+++ .github/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - [**Fixed**] `>colour` command always return incorrect value
 - [**Added**] Add loot table choice to `>barter` command
+- [**Added**] Add `>findblock` command
 
 ### Internal Changes
 - [**Added**] Add `ClampedRange` to revert some command's old range behaviour